django/requirements.txt
# Analyse et sécurité
yara-python==4.5.1
clamav==0.2
pyclamd==0.4.0
virustotal-api==1.1.11

# Django core
Django==4.2.20
djangorestframework==3.16.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3
mysql
# Sécurité
bcrypt==4.2.1
cryptography==44.0.0
PyNaCl==1.5.0
paramiko==3.5.0

# Utilitaires
psutil==7.0.0
pillow==11.1.0
reportlab==4.2.5

# Réseau/HTTP
requests==2.32.3
requests_ntlm==1.3.0
urllib3==2.3.0
chardet==5.2.0
charset-normalizer==3.4.1
certifi==2025.1.31
idna==3.10
xmltodict==0.14.2

# Autres
pyspnego==0.11.2
pywinrm==0.5.0
tzdata==2025.1